[develop] Add checks for correct SPP settings for the selected CCPP suite, resolve four more issues (#1259)
This PR adds a check in setup.py that ensures the appropriate physics schemes are being used for the selected SPP settings, and if not, removes the SPP settings for that scheme. If after this pruning there are no SPP settings left, an exception is raised. Note that many of the changes in setup.py are just re-arranging of existing code; the removed logic from lines 1029 to 1141 are moved downwards so that all of these CCPP checks are done only when a run_fcst task is specified in the workflow.

While making these changes to setup.py, I also added this file to the list of pytest-ed files. This required updating pylint to a newer version, since 2.17 results in an error (and this older version is no longer supported anyway).

Additionally, this PR resolves four more issues:
* Issue #1228, a new smoke/dust test smoke_dust_grid_RRFS_CONUS_25km_suite_RRFS_sas is added
* Issue #1258, removing smoke/dust entries from parm/field_table.RRFS_sas
* Issue #1275, fixing the order of operations error in setup.py after issue #1204 was merged
* Issue #1276, modified WE2E test get_from_HPSS_ics_GDAS_lbcs_GDAS_fmt_netcdf_2022040400_ensemble to test SPP_LSM, also change grid from CONUS 3km to CONUS 13km to make it much cheaper (we already have plenty CONUS 3km tests)

---------

Co-authored-by: Ben Koziol <bkoziol@redlineperf.com>
Co-authored-by: michael.lueken <Michael.Lueken@noaa.gov>